Adds IsLoading bindable property to pages and fixes SelectViewModelCommand (#45)
In Maui, the `IsBusy` property on pages now shows a loading indicator
with no way to override that behavior. Since implementing apps may want
their own loading indicator, this creates a new `IsLoading` property
which the implementing app can choose to forward to `IsBusy` if it
wants.

Also fixes the SelectViewModelCommand so it passes the model instead of
the view model to the `CanExecute` callback.